extract SearchResults component
