+++ frontend/src/metabase/nav/components/SearchBar.jsx
@@ -1,55 +1,15 @@
-/* eslint-disable react/prop-types */
 import React from "react";
+import PropTypes from "prop-types";
 import ReactDOM from "react-dom";
-import { Flex } from "grid-styled";
-import styled from "styled-components";
-import { space } from "styled-system";
 import { t } from "ttag";
-
-import { color, lighten } from "metabase/lib/colors";
-
-import Card from "metabase/components/Card";
 import Icon from "metabase/components/Icon";
 import OnClickOutsideWrapper from "metabase/components/OnClickOutsideWrapper";
-import SearchResult from "metabase/search/components/SearchResult";
-
-import { DefaultSearchColor } from "metabase/nav/constants";
 import MetabaseSettings from "metabase/lib/settings";
-
-const ActiveSearchColor = lighten(color("nav"), 0.1);
-
-import Search from "metabase/entities/search";
-
-const SearchWrapper = Flex.extend`
-  position: relative;
-  background-color: ${props =>
-    props.active ? ActiveSearchColor : DefaultSearchColor};
-  border-radius: 6px;
-  flex: 1 1 auto;
-  max-width: 50em;
-  align-items: center;
-  color: white;
-  transition: background 300ms ease-in;
-  &:hover {
-    background-color: ${ActiveSearchColor};
-  }
-`;
-
-const SearchInput = styled.input`
-  ${space};
-  background-color: transparent;
-  width: 100%;
-  border: none;
-  color: white;
-  font-size: 1em;
-  font-weight: 700;
-  &:focus {
-    outline: none;
-  }
-  &::placeholder {
-    color: ${color("text-white")};
-  }
-`;
+import SearchResults from "metabase/nav/components/SearchResults";
+import {
+  SearchWrapper,
+  SearchInput,
+} from "metabase/nav/components/SearchBar.styled.jsx";
 
 const ALLOWED_SEARCH_FOCUS_ELEMENTS = new Set(["BODY", "A"]);
 
@@ -94,25 +54,6 @@ export default class SearchBar extends React.Component {
     }
   };
 
-  renderResults(results) {
-    if (results.length === 0) {
-      return (
-        <li className="flex flex-column align-center justify-center p4 text-medium text-centered">
-          <div className="my3">
-            <Icon name="search" mb={1} size={24} />
-            <h3 className="text-light">{t`Didn't find anything`}</h3>
-          </div>
-        </li>
-      );
-    } else {
-      return results.map(l => (
-        <li key={`${l.model}:${l.id}`}>
-          <SearchResult result={l} compact={true} />
-        </li>
-      ));
-    }
-  }
-
   render() {
     const { active, searchText } = this.state;
     return (
@@ -146,23 +87,10 @@ export default class SearchBar extends React.Component {
           {active && MetabaseSettings.searchTypeaheadEnabled() && (
             <div className="absolute left right text-dark" style={{ top: 60 }}>
               {searchText.trim().length > 0 ? (
-                <Card
-                  className="overflow-y-auto"
-                  style={{ maxHeight: 400 }}
-                  py={1}
-                >
-                  <Search.ListLoader
-                    query={{ q: searchText.trim() }}
-                    wrapped
-                    reload
-                    debounced
-                  >
-                    {({ list }) => {
-                      return <ol>{this.renderResults(list)}</ol>;
-                    }}
-                  </Search.ListLoader>
-                </Card>
-              ) : null}
+                <SearchResults searchText={searchText} />
+              ) : (
+                <div>Recents!</div>
+              )}
             </div>
           )}
         </SearchWrapper>
@@ -170,3 +98,7 @@ export default class SearchBar extends React.Component {
     );
   }
 }
+SearchBar.propTypes = {
+  location: PropTypes.object,
+  onChangeLocation: PropTypes.func,
+};
